Verify a peer's standing at an intermediary to reduce double spending
Prior to this change there was no way for the client to verify a peer's current
standing at an intermediary. The only indication that a peer's ballance had
been exhausted was when a standing update was rejected. Moreover, the client
had no way of getting an updated state representation from an intermediary
after using it to download piece data. The meant even honest peers would
unavoidably attempt to double spend.

To close this hole two new messages are added: get_standing and your_standing.
These allow the client to verify a peer's standing at an intermediary and
inform the downloader of that state respectively.